# packages/shared/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/authorize/mcp/{serverIdentifier}/{version}": {
      "get": {
        "description": "Handles authorization requests from MCP clients. Validates the request, stores PKCE parameters, and redirects to the consent screen UI.",
        "operationId": "AuthorizationController_authorize",
        "parameters": [
          {
            "name": "response_type",
            "required": true,
            "in": "query",
            "description": "OAuth 2.0 response type (must be \"code\" for authorization code flow)",
            "schema": {
              "example": "code",
              "type": "string",
              "enum": [
                "code"
              ]
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "description": "Space-delimited list of scopes being requested",
            "schema": {
              "example": "tasks:read tasks:write",
              "type": "string"
            }
          },
          {
            "name": "client_id",
            "required": true,
            "in": "query",
            "description": "Client identifier issued during registration",
            "schema": {
              "example": "0bab273987a2e163c3abb40c631ec0a4",
              "type": "string"
            }
          },
          {
            "name": "code_challenge",
            "required": true,
            "in": "query",
            "description": "PKCE code challenge derived from the code verifier",
            "schema": {
              "example": "SVDkCojJ1edZzG47VFce6y2o5ieCQG19ueoYMQ_xloo",
              "type": "string"
            }
          },
          {
            "name": "code_challenge_method",
            "required": true,
            "in": "query",
            "description": "PKCE code challenge method (S256 for SHA-256)",
            "schema": {
              "example": "S256",
              "type": "string"
            }
          },
          {
            "name": "redirect_uri",
            "required": true,
            "in": "query",
            "description": "Redirect URI where the authorization response will be sent",
            "schema": {
              "example": "http://localhost:6274/oauth/callback/debug",
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "description": "Opaque state value for CSRF protection",
            "schema": {
              "example": "c5d4fa797d8c0a9025237aafb780ef7e9c38a721a00530372bc26c89572b1833",
              "type": "string"
            }
          },
          {
            "name": "resource",
            "required": true,
            "in": "query",
            "description": "Resource server URL that the client wants to access",
            "schema": {
              "example": "http://localhost:4001/",
              "type": "string"
            }
          },
          {
            "name": "serverIdentifier",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirects to consent screen with authorization request ID"
          },
          "400": {
            "description": "Invalid authorization request parameters"
          },
          "404": {
            "description": "MCP server or client not found"
          }
        },
        "summary": "OAuth 2.0 Authorization Endpoint",
        "tags": [
          "Authorization Server"
        ]
      },
      "post": {
        "description": "Handles user consent decision. Validates the flow ID (CSRF token), generates an authorization code if approved, and redirects back to the client with the code or error.",
        "operationId": "AuthorizationController_authorizeConsent",
        "parameters": [
          {
            "name": "serverIdentifier",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentDecisionDto"
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "Redirects to client redirect_uri with authorization code or error"
          },
          "400": {
            "description": "Invalid consent decision or flow state"
          },
          "401": {
            "description": "Authorization flow has already been used"
          },
          "404": {
            "description": "Authorization flow not found"
          }
        },
        "summary": "OAuth 2.0 Authorization Consent Handler",
        "tags": [
          "Authorization Server"
        ]
      }
    },
    "/api/v1/auth/flow/{flowId}": {
      "get": {
        "description": "Retrieves authorization flow details for the consent screen",
        "operationId": "AuthorizationController_getFlow",
        "parameters": [
          {
            "name": "flowId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorization flow details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpAuthorizationFlowEntity"
                }
              }
            }
          },
          "404": {
            "description": "Authorization flow not found"
          }
        },
        "summary": "Get authorization flow details",
        "tags": [
          "Authorization Server"
        ]
      }
    },
    "/api/v1/auth/token/mcp/{serverIdentifier}/{version}": {
      "post": {
        "description": "Exchanges authorization code for access token. Validates PKCE code_verifier, issues signed JWT access token and refresh token.",
        "operationId": "AuthorizationController_token",
        "parameters": [
          {
            "name": "serverIdentifier",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access token issued successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid token request parameters"
          },
          "401": {
            "description": "Invalid authorization code, code_verifier, or expired code"
          }
        },
        "summary": "OAuth 2.0 Token Endpoint",
        "tags": [
          "Authorization Server"
        ]
      }
    },
    "/api/v1/auth/introspect/mcp/{serverIdentifier}/{version}": {
      "post": {
        "description": "Introspects an access token to validate it and retrieve its metadata. Verifies JWT signature, expiration, and claims according to RFC 7662.",
        "operationId": "AuthorizationController_introspect",
        "parameters": [
          {
            "name": "serverIdentifier",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntrospectTokenRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token introspection response (active true/false with metadata)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntrospectTokenResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid introspection request parameters"
          }
        },
        "summary": "OAuth 2.0 Token Introspection Endpoint",
        "tags": [
          "Authorization Server"
        ]
      }
    },
    "/api/v1/auth/token-exchange/mcp/{serverIdentifier}/{version}": {
      "post": {
        "description": "Exchanges MCP JWT access token for downstream system tokens. Validates MCP token, resolves scope mappings, and returns downstream access token with automatic refresh if needed.",
        "operationId": "AuthorizationController_tokenExchange",
        "parameters": [
          {
            "name": "serverIdentifier",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenExchangeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token exchange successful - returns downstream access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenExchangeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid token exchange request parameters"
          },
          "401": {
            "description": "Invalid or expired MCP JWT"
          },
          "403": {
            "description": "Insufficient scope - requested scope not entitled"
          },
          "404": {
            "description": "Connection not found"
          }
        },
        "summary": "RFC 8693 Token Exchange Endpoint",
        "tags": [
          "Authorization Server"
        ]
      }
    },
    "/api/v1/auth/callback": {
      "get": {
        "description": "Handles callbacks from downstream OAuth providers. Validates the state, exchanges authorization code for tokens, and continues the auth flow.",
        "operationId": "AuthorizationController_callback",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "description": "Authorization code from downstream OAuth provider",
            "schema": {
              "example": "abc123xyz",
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "description": "State parameter that identifies the connection flow",
            "schema": {
              "example": "xyz789abc",
              "type": "string"
            }
          },
          {
            "name": "error",
            "required": false,
            "in": "query",
            "description": "Error code if authorization failed",
            "schema": {
              "example": "access_denied",
              "type": "string"
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "description": "Space-delimited list of scopes that were granted",
            "schema": {
              "example": "tasks:read tasks:write",
              "type": "string"
            }
          },
          {
            "name": "error_description",
            "required": false,
            "in": "query",
            "description": "Error description if authorization failed",
            "schema": {
              "example": "User denied the authorization request",
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirects to next step in the authorization flow"
          },
          "400": {
            "description": "Invalid callback parameters or state"
          },
          "404": {
            "description": "Connection flow not found for provided state"
          }
        },
        "summary": "OAuth 2.0 Callback Endpoint for Downstream Systems",
        "tags": [
          "Authorization Server"
        ]
      }
    },
...
  },
  "info": {
    "title": "AI Monorepo API",
    "description": "The AI Monorepo API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-Cookie": {
        "type": "http",
        "in": "cookie",
        "name": "access_token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ConsentDecisionDto": {
        "type": "object",
        "properties": {
          "flow_id": {
            "type": "string",
            "description": "Authorization flow ID (serves as CSRF protection token)",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "approved": {
            "type": "boolean",
            "description": "Whether the user approved the authorization request",
            "example": true
          }
        },
        "required": [
          "flow_id",
          "approved"
        ]
      },
      "McpAuthorizationFlowEntity": {
        "type": "object",
        "properties": {}
      },
      "TokenRequestDto": {
        "type": "object",
        "properties": {
          "grant_type": {
            "type": "string",
            "description": "Grant type that determines which parameters must be supplied",
            "enum": [
              "authorization_code",
              "refresh_token"
            ],
            "example": "authorization_code"
          },
          "client_id": {
            "type": "string",
            "description": "Client identifier issued during dynamic registration",
            "example": "0bab273987a2e163c3abb40c631ec0a4"
          },
          "code": {
            "type": "string",
            "description": "Authorization code that was issued by the /authorize endpoint",
            "example": "SplxlOBeZQQYbYS6WxSbIA"
          },
          "redirect_uri": {
            "type": "string",
            "description": "Redirect URI used during authorization (required when code is present)",
            "example": "http://localhost:6274/oauth/callback/debug"
          },
          "code_verifier": {
            "type": "string",
            "description": "PKCE code verifier used to validate the authorization code exchange",
            "example": "dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk"
          },
          "refresh_token": {
            "type": "string",
            "description": "Refresh token issued earlier by the authorization server",
            "example": "def5020091c58c49fbb..."
          },
          "scope": {
            "type": "string",
            "description": "Optional list of scopes to narrow when refreshing a token (space-delimited)",
            "example": "tasks:read tasks:write"
          },
          "resource": {
            "type": "string",
            "description": "Optional resource indicator (RFC 8707) - identifies the target resource server",
            "example": "http://localhost:4001/"
          }
        },
        "required": [
          "grant_type",
          "client_id"
        ]
      },
      "TokenResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "Opaque access token used to access protected resources",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "token_type": {
            "type": "string",
            "description": "Type of token issued (MCP clients always receive Bearer tokens)",
            "enum": [
              "Bearer"
            ],
            "example": "Bearer",
            "default": "Bearer"
          },
          "expires_in": {
            "type": "number",
            "description": "Lifetime of the access token in seconds",
            "example": 3600
          },
          "refresh_token": {
            "type": "string",
            "description": "Refresh token that can be exchanged for a new access token",
            "example": "def5020091c58c49fbb..."
          },
          "scope": {
            "type": "string",
            "description": "Space-delimited scopes that were granted for this token",
            "example": "tasks:read tasks:write"
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "refresh_token"
        ]
      },
      "IntrospectTokenRequestDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Access or refresh token that should be validated",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "token_type_hint": {
            "type": "string",
            "description": "Hint to help the server determine the token lookup strategy",
            "enum": [
              "access_token",
              "refresh_token"
            ],
            "example": "access_token"
          },
          "client_id": {
            "type": "string",
            "description": "Client identifier for optional validation against the token claims. Per RFC 7662, this is not required - the client_id is extracted from the token itself.",
            "example": "0bab273987a2e163c3abb40c631ec0a4"
          },
          "client_secret": {
            "type": "string",
            "description": "Client secret for confidential clients (MCP clients typically omit this)",
            "example": "s3cr3t",
            "nullable": true
          }
        },
        "required": [
          "token"
        ]
      },
      "IntrospectTokenResponseDto": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "description": "Indicates whether the token is currently valid",
            "example": true
          },
          "token_type": {
            "type": "string",
            "description": "Token type, always Bearer for MCP clients",
            "enum": [
              "Bearer"
            ],
            "example": "Bearer"
          },
          "client_id": {
            "type": "string",
            "description": "Client identifier associated with the token",
            "example": "0bab273987a2e163c3abb40c631ec0a4"
          },
          "sub": {
            "type": "object",
            "description": "Subject of the token (resource owner or actor)",
            "example": "journey:1234"
          },
          "aud": {
            "description": "Audience that should accept this token",
            "oneOf": [
              {
                "type": "string",
                "example": "taskeroo-api"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "taskeroo-api"
                ]
              }
            ]
          },
          "iss": {
            "type": "object",
            "description": "Issuer that minted the token",
            "example": "https://auth.taskeroo.local/auth"
          },
          "jti": {
            "type": "object",
            "description": "Unique token identifier for replay detection",
            "example": "b15e8a76-5b6d-4bde-9a3b-26fdbaab5b4c"
          },
          "exp": {
            "type": "object",
            "description": "Expiration timestamp (seconds since Unix epoch)",
            "example": 1731145219
          },
          "iat": {
            "type": "object",
            "description": "Issued-at timestamp (seconds since Unix epoch)",
            "example": 1731141619
          },
          "nbf": {
            "type": "number",
            "description": "Not-before timestamp (seconds since Unix epoch)",
            "example": 1731141019
          },
          "scope": {
            "type": "string",
            "description": "Granted scopes (space-delimited) for display purposes",
            "example": "tasks:read tasks:write"
          },
          "server_identifier": {
            "type": "object",
            "description": "MCP server identifier the token is scoped to",
            "example": "taskeroo"
          },
          "resource": {
            "type": "object",
            "description": "Resource URL that was used during authorization",
            "example": "http://localhost:4001/"
          },
          "version": {
            "type": "object",
            "description": "Version of the MCP server contract",
            "example": "1.0.0"
          }
        },
        "required": [
          "active",
          "token_type",
          "client_id",
          "sub",
          "aud",
          "iss",
          "jti",
          "exp",
          "iat",
          "server_identifier",
          "resource",
          "version"
        ]
      },
      "TokenExchangeRequestDto": {
        "type": "object",
        "properties": {
          "grant_type": {
            "type": "string",
            "description": "Grant type for token exchange as defined in RFC 8693",
            "enum": [
              "urn:ietf:params:oauth:grant-type:token-exchange"
            ],
            "example": "urn:ietf:params:oauth:grant-type:token-exchange"
          },
          "subject_token": {
            "type": "string",
            "description": "The access token that represents the identity of the party on behalf of whom the request is being made",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "subject_token_type": {
            "type": "string",
            "description": "Type identifier for the subject_token as defined in RFC 8693",
            "enum": [
              "urn:ietf:params:oauth:token-type:access_token"
            ],
            "example": "urn:ietf:params:oauth:token-type:access_token"
          },
          "resource": {
            "type": "string",
            "description": "Resource server URL that the client wants to access with the exchanged token",
            "example": "http://localhost:4001/"
          },
          "scope": {
            "type": "string",
            "description": "Optional space-delimited list of scopes for the exchanged token",
            "example": "tasks:read tasks:write"
          }
        },
        "required": [
          "grant_type",
          "subject_token",
          "subject_token_type",
          "resource"
        ]
      },
      "TokenExchangeResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "The newly issued access token after successful exchange",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "issued_token_type": {
            "type": "string",
            "description": "Type identifier for the issued token as defined in RFC 8693",
            "example": "urn:ietf:params:oauth:token-type:access_token"
          },
          "token_type": {
            "type": "string",
            "description": "Type of token issued, always Bearer for MCP clients",
            "example": "Bearer"
          },
          "expires_in": {
            "type": "number",
            "description": "Lifetime of the access token in seconds",
            "example": 3600
          },
          "scope": {
            "type": "string",
            "description": "Space-delimited list of scopes granted for the exchanged token",
            "example": "tasks:read tasks:write"
          }
        },
        "required": [
          "access_token",
          "issued_token_type",
          "token_type",
          "expires_in",
          "scope"
        ]
      },
...
    }
  }
}